`Communication`: Migrate ConversationList to Observable (#267)
* Migrate ConversationList to Observable

* Refactor

* Use ObservationTracking for filter changes

* Fix missing view updates

* Move Filter enum

* Move search into viewModel
